design_parser_program/requirements.txt
# Design Parser Requirements
# Install with: pip install -r requirements.txt

# Core image processing
Pillow>=9.0.0

# Design file conversion tools (optional)
fig2sketch[fast]>=0.0.1
sketch-tool>=0.0.1

# 3D file processing (optional)
psd-tools>=1.9.0

# Video processing (optional)
ffmpeg-python>=0.2.0

# Additional utilities
python-magic>=0.4.27
chardet>=5.0.0

# Optional: For advanced image analysis
opencv-python>=4.5.0
numpy>=1.21.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

